Move constants to their own module
